Support for non-table filtering
Changed all filtering code to not depend on jQuery.
Tables must now manually specify the filterable identifier field in the header cell, it is no longer assumed from the text of the cell.
Added !important to CSS that hides elements that are filtered/paged, which is needed due to CSS priority rules.
README still needs to be updated for these changes.