Filter buttons upgrade
Dispatch change event when filters are reset and cleared, so any other scripts can be notifed that they were changed.
Added "noELC" detail so ELC can be told to ignore an event, as in the above case.
Use call() instead of dispatchEvent() for internal events in cases where we don't need to notify all scripts that something is happening.