Version Updates
Taking the risk that no one is actualy using the former v1.0.0. I renamed it to v0.0 which seems more appropriate for how archaic it is.
Naming the current version v0.1, because I'm using it in various projects and don't want to break them with future changes all at once.